requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
cryptography==42.0.2
dataclasses-json==0.6.3
defusedxml==0.8.0rc2
distro==1.8.0
Django==4.2.7
django-autocomplete-light==3.9.7
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djhacker==0.2.3
djoser==2.2.2
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langsmith==0.0.70
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.4.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pluggy==1.4.0
protobuf==4.25.1
pyarrow==14.0.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyhOn==0.15.14
PyJWT==2.8.0
pypdf==3.17.2
PyPDF2==3.0.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
SQLAlchemy==2.0.23
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
yarl==1.9.4
zipp==3.17.0
